QPR-13958 fixes

+++ OREAnalytics/orea/engine/valuationengine.hpp
@@ -115,7 +115,7 @@ public:
         
 private:
     void recalibrateModels();
-    std::tuple<double, double, double, double>
+    std::tuple<long, long, long, long>
     populateCube(const QuantLib::Date& d, size_t cubeDateIndex, size_t sample, bool isValueDate, bool isStickyDate,
                  bool scenarioUpdated, const std::map<std::string, QuantLib::ext::shared_ptr<ore::data::Trade>>& trades,
                  const std::vector<QuantLib::ext::shared_ptr<ore::data::OptionWrapper>>& optionWrappers,
